Added tasks 3020-3046

+++ src/main/kotlin/g2801_2900/s2815_max_pair_sum_in_an_array/Solution.kt
@@ -23,7 +23,7 @@ class Solution {
                 continue
             }
             val sum = value.poll() + value.poll()
-            maxSum = max(maxSum.toDouble(), sum.toDouble()).toInt()
+            maxSum = max(maxSum, sum)
         }
         return maxSum
     }
